+++ components/pages/MyPage.tsx
@@ -2,7 +2,10 @@ import React from "react";
 import { View, Text, StyleSheet, Image } from "react-native";
 import Icon from "react-native-vector-icons/FontAwesome";
 import SafeAreaView from "react-native-safe-area-view";
-import { ScrollView } from "react-native-gesture-handler";
+import { ScrollView, TouchableOpacity } from "react-native-gesture-handler";
+import { createStackNavigator } from "@react-navigation/stack";
+import SignUp from "./SignUp";
+import { useNavigation } from "@react-navigation/native";
 
 const styles = StyleSheet.create({
   container: {
@@ -59,60 +62,67 @@ const styles = StyleSheet.create({
   },
 });
 
-export default function MyPage() {
+const Stack = createStackNavigator();
+
+function MyPage() {
+  const { navigate } = useNavigation();
+
   return (
     <SafeAreaView>
       <ScrollView>
-        <View style={styles.container}>
+        <TouchableOpacity
+          style={styles.container}
+          onPress={() => navigate("SignUp")}
+        >
           <Image
             source={require("../../assets/user_icon.png")}
             style={styles.icon}
           />
           <Text style={styles.headerTitle}>会員情報・ログインへ</Text>
-        </View>
+        </TouchableOpacity>
         <View>
           <View style={styles.boxes}>
-            <View style={styles.box}>
+            <TouchableOpacity style={styles.box}>
               <Text>いいね！・閲覧履歴</Text>
               <Icon name="angle-right" size={24} color="#ccc" />
-            </View>
-            <View style={styles.box}>
+            </TouchableOpacity>
+            <TouchableOpacity style={styles.box}>
               <Text>保存した検索条件</Text>
               <Icon name="angle-right" size={24} color="#ccc" />
-            </View>
+            </TouchableOpacity>
           </View>
           <View style={styles.boxes}>
-            <View style={styles.box}>
+            <TouchableOpacity style={styles.box}>
               <Text>出品した商品</Text>
               <Icon name="angle-right" size={24} color="#ccc" />
-            </View>
-            <View style={styles.box}>
+            </TouchableOpacity>
+            <TouchableOpacity style={styles.box}>
               <Text>購入した商品</Text>
               <Icon name="angle-right" size={24} color="#ccc" />
-            </View>
-            <View style={styles.box}>
+            </TouchableOpacity>
+            <TouchableOpacity style={styles.box}>
               <Text>下書き一覧</Text>
               <Icon name="angle-right" size={24} color="#ccc" />
-            </View>
-            <View style={styles.box}>
+            </TouchableOpacity>
+            <TouchableOpacity style={styles.box}>
               <Text>持ち物一覧</Text>
               <Icon name="angle-right" size={24} color="#ccc" />
-            </View>
+            </TouchableOpacity>
           </View>
           <View style={styles.boxes}>
             <Text style={styles.boxTitle}>ガイド・お問い合わせ</Text>
-            <View style={styles.box}>
+            <TouchableOpacity style={styles.box}>
               <Text>ガイド</Text>
               <Icon name="angle-right" size={24} color="#ccc" />
-            </View>
-            <View style={styles.box}>
+            </TouchableOpacity>
+            <TouchableOpacity style={styles.box}>
               <Text>お問い合わせ</Text>
               <Icon name="angle-right" size={24} color="#ccc" />
-            </View>
-            <View style={styles.box}>
+            </TouchableOpacity>
+            <TouchableOpacity style={styles.box}>
               <Text>メルカリボックス</Text>
               <Icon name="angle-right" size={24} color="#ccc" />
-            </View>
+            </TouchableOpacity>
           </View>
           <View style={styles.boxes}>
             <View style={styles.box}>
@@ -121,13 +131,34 @@ export default function MyPage() {
             </View>
           </View>
           <View style={styles.boxes}>
-            <View style={styles.box}>
+            <TouchableOpacity style={styles.box}>
               <Text>キャッシュを消去する</Text>
               <Icon name="angle-right" size={24} color="#ccc" />
-            </View>
+            </TouchableOpacity>
           </View>
         </View>
       </ScrollView>
     </SafeAreaView>
+  );
+}
+
+export default function () {
+  return (
+    <Stack.Navigator mode="modal">
+      <Stack.Screen
+        name="MyPage"
+        component={MyPage}
+        options={{
+          headerShown: false,
+        }}
+      />
+      <Stack.Screen
+        name="SignUp"
+        component={SignUp}
+        options={{
+          headerShown: false,
+        }}
+      />
+    </Stack.Navigator>
   );
 }
