+++ components/pages/SignIn.tsx
@@ -3,9 +3,6 @@ import { Image, StyleSheet, Text, View } from "react-native";
 import { TouchableOpacity } from "react-native-gesture-handler";
 import SafeAreaView from "react-native-safe-area-view";
 import Icon from "react-native-vector-icons/FontAwesome";
-import { createStackNavigator } from "@react-navigation/stack";
-import SignIn from "./SignIn";
-import { useNavigation } from "@react-navigation/native";
 
 const styles = StyleSheet.create({
   component: {
@@ -25,16 +22,8 @@ const styles = StyleSheet.create({
   },
   boxes: {
     alignItems: "center",
-    borderColor: "black",
-    borderTopWidth: 1,
-    marginTop: 30,
+    marginTop: 20,
     paddingTop: 15,
-  },
-  boxTitle: {
-    backgroundColor: "white",
-    position: "absolute",
-    top: -16,
-    padding: 10,
   },
   box: {
     width: 300,
@@ -66,21 +55,9 @@ const styles = StyleSheet.create({
     color: "black",
     fontWeight: "bold",
   },
-  login: {
-    marginTop: 20,
-    borderBottomColor: "#136FFF",
-    borderBottomWidth: 1,
-  },
-  loginText: {
-    color: "#136FFF",
-  },
 });
 
-const Stack = createStackNavigator();
-
-function SignUp() {
-  const { navigate } = useNavigation();
-
+export default function SignIn() {
   return (
     <SafeAreaView style={styles.component}>
       <View style={styles.title}>
@@ -91,58 +68,29 @@ function SignUp() {
         <Text style={styles.titleText}>mercari</Text>
       </View>
       <View style={styles.boxes}>
-        <Text style={styles.boxTitle}>初めてご利用の方は</Text>
         <View>
           <TouchableOpacity style={[styles.box, styles.blackBackgroundColor]}>
             <Icon name="apple" size="24" color="white" />
-            <Text style={styles.whiteColor}>Appleでサインアップ</Text>
+            <Text style={styles.whiteColor}>Appleでログイン</Text>
             <Text></Text>
           </TouchableOpacity>
           <TouchableOpacity style={[styles.box, styles.blueBackgroundColor]}>
             <Icon name="facebook-official" size="24" color="white" />
-            <Text style={styles.whiteColor}>Facebookで登録</Text>
+            <Text style={styles.whiteColor}>Facebookでログイン</Text>
             <Text></Text>
           </TouchableOpacity>
           <TouchableOpacity style={[styles.box, styles.whiteBackgroundColor]}>
             <Icon name="google" size="24" />
-            <Text style={styles.blackColor}>Googleで登録</Text>
+            <Text style={styles.blackColor}>Googleでログイン</Text>
             <Text></Text>
           </TouchableOpacity>
           <TouchableOpacity style={[styles.box, styles.redBackgroundColor]}>
             <Icon name="envelope-o" size="24" color="white" />
-            <Text style={styles.whiteColor}>メールアドレスで登録</Text>
+            <Text style={styles.whiteColor}>メールアドレスでログイン</Text>
             <Text></Text>
           </TouchableOpacity>
         </View>
       </View>
-      <View style={styles.login}>
-        <TouchableOpacity onPress={() => navigate("SignIn")}>
-          <Text style={styles.loginText}>こちらからログイン</Text>
-        </TouchableOpacity>
-      </View>
     </SafeAreaView>
-  );
-}
-
-export default function () {
-  return (
-    <Stack.Navigator mode="card">
-      <Stack.Screen
-        name="MyPage"
-        component={SignUp}
-        options={{
-          title: "会員登録",
-          headerBackTitle: "",
-        }}
-      />
-      <Stack.Screen
-        name="SignIn"
-        component={SignIn}
-        options={{
-          title: "ログイン",
-          headerBackTitleVisible: false,
-        }}
-      />
-    </Stack.Navigator>
   );
 }
